Fixed test failures in v0.10 and v0.12 by moving object shorthand tests to a file that is only tested in node v4 and up.
